Rework the plugin to get a REQUEST param
