--- plugins/layercustomproperties/__init__.py
@@ -1,4 +1,0 @@
-def serverClassFactory(serverIface):
-    from . layercustomproperties import GetLayerCustomProperties
-    return GetLayerCustomProperties(serverIface)
-
