+++ plugins/openquakeqgiswebservices/README.md
