+++ plugins/openquakeqgiswebservices/__init__.py
@@ -1,0 +1,4 @@
+def serverClassFactory(serverIface):
+    from . openquakeqgiswebservices import OpenQuakeQGIS
+    return OpenQuakeQGIS(serverIface)
+
