+++ plugins/openquakeqgiswebservices/metadata.txt
